+++ src/RequestLocation/FormParamLocation.php
@@ -68,7 +68,7 @@ class FormParamLocation extends AbstractLocation
             }
         }
 
-        $body = http_build_query(NonFiniteFloats::normalizeAll($data['form_params']), '', '&');
+        $body = http_build_query(NonFiniteFloats::normalizeAll($data['form_params'], 'a formParam location value'), '', '&');
         $modify['body'] = Psr7\Utils::streamFor($body);
         $modify['set_headers']['Content-Type'] = $this->contentType;